Add files via upload
CHANGES DONE

1. Asking about the Ritual
Made slightly more random.
The order now goes:
	Not Role 1 decided by pre-shuffle of the roles
	Not Role 3 decided by pre-shuffle of the roles
	Not Role 4 decided by pre-shuffle of the roles
	Then random of
		Two thirds chance of Not Role 1 again
		One third chance of Not Role 4 again

2. Giving Sigil to Role 1
Calls up EventEliminated so will use the same image

3. Initialisation for 5 Investigators
Added for game balance.

4. Close Friend Text added for asking about Role 3
If an Ally is asked about Ally Role 3 they will reveal the Ally was a close friend of Martha.  For roles 1 and 2, this may be a lie.  Once done, a flag is raised so that the text is not repeated.

5. Darkness Tokens
Darkness Tokens cover all spaces of a tile.
Text added to Scenario and to the placement of tiles.

6. Cellar Crate
Added test for the searching of the crate

7. Ghost of Bettina
Fixed reference to husband and step-sister.